fixed crash due to double template deletion
Template consistency ensured outside of this scope, e.g. when a part of
a template is removed, the according part of the code needs to take care
that prev/next/first/last pointers are all still valid or be
invalidated.

A single template engine always consists of only one object, i.e.
articulated vehicles and such are resembled as a single template object.