fixed crash during template replacement
This occurred only sometimes. What happened was that the _new_vehicle_id
sometimes was still 0 and sometimes not. In case of 0 the chance was
there that this was actually a valid ID of an existing vehicle.

If this ID happened to be !=0 it was likely that no vehicle of this ID
could be found, resulting in a nullptr in new_vehicle in line 358.

Now we excluded these steps from the flags==DC_NONE case with an
additional check. This means that new_vehicle might still be NULL during
the refit check below.